tasks delete works but now list delete does not

+++ app/Controllers/ListsController.js
@@ -1,5 +1,6 @@
 import { ProxyState } from "../AppState.js"
 import { listsService } from "../Services/ListsService.js"
+import { tasksService } from "../Services/TasksService.js"
 
 
 
@@ -17,8 +18,10 @@ function _draw() {
 //PUBLIC
 export default class ListsController {
   constructor() {
+
     ProxyState.on('lists', _draw)
     ProxyState.on('tasks', _draw)
+    ProxyState.on('taskCount', _draw)
     _draw()
   }
 
@@ -27,14 +30,17 @@ export default class ListsController {
     let form = window.event.target
     let listElement = {
       name: form['name'].value,
-      color: form['color'].value
+      color: form['color'].value,
     }
     listsService.submitList(listElement)
     console.log("submitList button from ListsController")
+    form.reset()
   }
 
   deleteList(listId) {
-    listsService.deleteList(listId)
+    if (window.confirm("Are You Sure You Want To Delete This List?")) {
+      listsService.deleteList(listId);
+    }
   }
